requirements.txt
numpy==2.0.1
pandas==2.2.2
fastapi==0.115.0
uvicorn==0.30.1
pillow==10.3.0
requests~=2.32.5
langgraph~=0.6.10
python-dotenv~=1.0.1
python-multipart~=0.0.9
prometheus-fastapi-instrumentator

# Evaluation & embeddings
sentence-transformers>=3.0.0
scikit-learn>=1.4.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Development tools
black==25.9.0
isort==7.0.0
mypy==1.18.2
flake8==7.3.0
pre-commit==3.6.2

# Additional
minio==7.1.6  # For MinIO client support
redis==4.0.2  # Redis client for Python
rq
pymongo==4.6.3  # MongoDB client for Python
qdrant-client==1.9.0
